feat: Add getCurriculaByMajorID function
This commit adds a new function, getCurriculaByMajorID, to the curricula/queries.ts file. This function retrieves the curricula data for a specific major ID. It makes an API call to fetch the data and returns a Promise with the response.

The function is intended to be used in the application to retrieve curricula data based on the selected major. It will be used in the PlanSettingPopup component to fetch the curricula for the selected major.

This commit follows the established commit message conventions in the repository.